Fixed some integer calculations which were breaking the conversion code.

+++ nImO/nImOarray.cpp
@@ -631,7 +631,7 @@ const
         std::queue<double> doublesSeen;
 
         outMessage.appendBytes(&startArray, sizeof(startArray));
-        writeInt64ToMessage(outMessage, inherited2::size() + DataKindIntegerShortValueMinValue - 1);
+        writeInt64ToMessage(outMessage, static_cast<int>(inherited2::size()) + DataKindIntegerShortValueMinValue - 1);
         for (const_iterator walker(inherited2::begin()); inherited2::end() != walker; ++walker)
         {
             SpValue aValue(*walker);
